Move withoutNulls() to util.
